+++ 1.20.1/src/main/java/com/torr/materia/client/ClientRenderLayers.java
@@ -4,6 +4,7 @@ import com.torr.materia.materia;
 import com.torr.materia.client.model.CartBodyModel;
 import com.torr.materia.client.model.CartCoverModel;
 import com.torr.materia.client.model.CartLanternModel;
+import com.torr.materia.client.model.CartShieldModel;
 import com.torr.materia.client.model.CartModel;
 import com.torr.materia.client.renderer.CustomSheepFurLayer;
 import net.minecraft.client.renderer.entity.SheepRenderer;
@@ -22,6 +23,7 @@ public class ClientRenderLayers {
         event.registerLayerDefinition(CartBodyModel.LAYER, CartBodyModel::createBodyLayer);
         event.registerLayerDefinition(CartCoverModel.LAYER, CartCoverModel::createBodyLayer);
         event.registerLayerDefinition(CartLanternModel.LAYER, CartLanternModel::createBodyLayer);
+        event.registerLayerDefinition(CartShieldModel.LAYER, CartShieldModel::createBodyLayer);
     }
     
     @SubscribeEvent